// pocketIso/src/administration-panel/models/organizational-context.model.ts
export interface OrganizationalContext {
    id: string;
    industry: string;
    factorsTechnologicalEnvironment: string;
    factorsRelatedToValuesCulture: string;
    externalAndInternalEntities: string;
    strengths: string;
    weaknesses: string;
    chances: string;
    threats: string;
    norm: number;
    numberNorm: string;
    normJustification: string;
  }